[PB-1525] add handleWebsiteLanguage service
